Consolidate list remote users api
