+++ kolibri/core/discovery/api.py
@@ -77,6 +77,7 @@ class StaticNetworkLocationViewSet(NetworkLocationViewSet):
 
 class NetworkLocationFacilitiesView(BaseValuesViewset):
     queryset = NetworkLocation.objects.all()
+    values = ()
     permission_classes = [NetworkLocationPermissions | NotProvisionedHasPermission]
     values = ("device_id", "instance_id", "device_name", "device_address", "facilities")
