+++ kolibri/plugins/setup_wizard/api.py
@@ -2,7 +2,6 @@ from django.urls import reverse
 from django.utils.decorators import method_decorator
 from django.views.decorators.csrf import csrf_protect
 from rest_framework import decorators
-from rest_framework.exceptions import AuthenticationFailed
 from rest_framework.exceptions import NotFound
 from rest_framework.exceptions import PermissionDenied
 from rest_framework.exceptions import ValidationError
@@ -14,7 +13,6 @@ from kolibri.core.auth.backends import FACILITY_CREDENTIAL_KEY
 from kolibri.core.auth.constants import user_kinds
 from kolibri.core.auth.models import Facility
 from kolibri.core.auth.models import FacilityUser
-from kolibri.core.auth.utils.users import get_remote_users_info
 from kolibri.core.device.models import DevicePermissions
 from kolibri.core.discovery.utils.network.client import NetworkClient
 from kolibri.core.discovery.utils.network.errors import NetworkLocationResponseFailure
@@ -145,33 +143,3 @@ class FacilityImportViewSet(ViewSet):
 
         except ValidationError:
             raise ValidationError(detail="duplicate", code="duplicate_username")
-
-    @decorators.action(methods=["post"], detail=False)
-    def listfacilitylearners(self, request):
-        """
-        If the request is done by an admin user  it will return a list of the users of the
-        facility
-
-        :param baseurl: First part of the url of the server that's going to be requested
-        :param facility_id: Id of the facility to authenticate and get the list of users
-        :param username: Username of the user that's going to authenticate
-        :param password: Password of the user that's going to authenticate
-        :return: List of the learners of the facility.
-        """
-        facility_id = request.data.get("facility_id")
-        baseurl = request.data.get("baseurl")
-        password = request.data.get("password")
-        username = request.data.get("username")
-        try:
-            facility_info = get_remote_users_info(
-                baseurl, facility_id, username, password
-            )
-        except AuthenticationFailed:
-            raise PermissionDenied()
-        user_info = facility_info["user"]
-        roles = user_info["roles"]
-        admin_roles = (user_kinds.ADMIN, user_kinds.SUPERUSER)
-        if not any(role in roles for role in admin_roles):
-            raise PermissionDenied()
-        students = [u for u in facility_info["users"] if not u["roles"]]
-        return Response({"students": students, "admin": facility_info["user"]})
