+++ kolibri/plugins/user_profile/assets/src/composables/useRemoteFacility.js
@@ -4,22 +4,22 @@ import urls from 'kolibri.urls';
 function remoteFacilityUserData(baseurl, facility_id, username, password, userAdmin = null) {
   const params = {
     baseurl: baseurl,
-    facility: facility_id,
+    facility_id: facility_id,
     username: userAdmin === null ? username : userAdmin,
     password: password,
   };
   return client({
-    url: urls['kolibri:kolibri.plugins.user_profile:remotefacilityauthenticateduserinfo'](),
+    url: urls['kolibri:core:remotefacilityauthenticateduserinfo'](),
     method: 'POST',
     data: params,
-  }).then(response => {
-    if (response.data.error) {
-      return 'error';
-    } else {
+  })
+    .then(response => {
       const user_info = response.data.find(element => element.username === username);
       return user_info;
-    }
-  });
+    })
+    .catch(() => {
+      return 'error';
+    });
 }
 
 const remoteFacilityUsers = function (baseurl, facility_id, username) {
@@ -29,7 +29,7 @@ const remoteFacilityUsers = function (baseurl, facility_id, username) {
     username: username,
   };
   return client({
-    url: urls['kolibri:kolibri.plugins.user_profile:remotefacilityuser'](),
+    url: urls['kolibri:core:remotefacilityuser'](),
     params: params,
   }).then(response => {
     let users = response.data;
